Playstore Assignment
The excercise was easy for me in regards to knowing which headers to pass through.
First I created the directory, installed Node, and created App.js

Then I installed nodemon, and morgan. Then I first wrote the validations needed for the headers that come through.

After writing out the validations, I sorted the results based on Genre and displayed that in a response.json().

I copied the sort from the previous excercise.